pyproject.toml
[tool.poetry]
name = "wisup_e2m"
version = "0.1.1"
description = "Everything to Markdown."
authors = ["Wisup Team <team@wisup.a>"]
license = "MIT"
repository = "https://github.com/wisupai/e2m"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
litellm = "^1.42.12"
pydantic = "^2.8.2"
tomlkit = "0.12.0"
surya-ocr = "^0.4.15"
torch = "^2.4.0"
marker-pdf = "^0.2.16"
unstructured = { extras = ["all"], version = "^0.15.0" }
pdfminer-six = "^20240706"
pillow-heif = "^0.18.0"
matplotlib = "^3.9.1"
unstructured-inference = "^0.7.36"
unstructured-pytesseract = "^0.3.12"
pikepdf = "^9.1.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
isort = "^5.13.2"
pytest = "^8.3.2"
lint = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
